Reorganize repository to follow Go project layout standards
- Move main.go to /cmd/concourse-replicated-channel-resource/
- Split resource.go into modular files in /internal/resource/
  - types.go: All type definitions and client code
  - check.go: Check operation implementation
  - in.go: In operation implementation
  - out.go: Out operation implementation
- Move build files to /build/ directory
- Move CI configuration to /deployments/ci/
- Update package declarations and imports
- Update build configurations for new structure
- All tests passing and binary compiles successfully
