+++ deployments/ci/set-pipeline.sh
