+++ internal/resource/in_test.go
@@ -1,4 +1,4 @@
-package main
+package resource
 
 import (
 	"context"
